requirements.txt
langchain==0.1.20
langchain_huggingface==0.0.3
transformers==4.41.0
sentencepiece>=0.1.99,<0.3.0
torch>=2.0.0,<2.4.0
torchvision>=0.15.0,<0.19.0
psycopg2-binary>=2.9.0,<2.10.0
numpy<2.0.0
pgvector>=0.1.8,<0.3.0
accelerate>=0.25.0,<0.32.0
urllib3<2.0
langchain-community~=0.0.37
pandas==2.2.2
tqdm~=4.67.1
scikit-learn~=1.6.1